note: Rename note driver to noteram driver
to reflect this is just a normal note driver with ram buffer
