+++ boards/arm/stm32f7/stm32f746-ws/configs/nsh/defconfig
@@ -32,6 +32,7 @@ CONFIG_CDCACM=y
 CONFIG_CDCACM_EPINTIN_HSSIZE=512
 CONFIG_DEBUG_SYMBOLS=y
 CONFIG_DRIVER_NOTE=y
+CONFIG_DRIVER_NOTERAM_BUFSIZE=512
 CONFIG_FAT_DMAMEMORY=y
 CONFIG_FAT_LCNAMES=y
 CONFIG_FAT_LFN=y
@@ -71,7 +72,6 @@ CONFIG_SCHED_HPWORKSTACKSIZE=1800
 CONFIG_SCHED_INSTRUMENTATION=y
 CONFIG_SCHED_LPWORK=y
 CONFIG_SCHED_LPWORKSTACKSIZE=1800
-CONFIG_SCHED_NOTE_BUFSIZE=512
 CONFIG_SCHED_WAITPID=y
 CONFIG_SDCLONE_DISABLE=y
 CONFIG_SIG_SIGWORK=4
